# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sabaik Almasa API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pages/{slug}:
    get:
      operationId: getSeoPage
      tags: [content]
      summary: Get a published SEO landing page
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: SEO landing page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeoPage"
        "404":
          description: Not found
  /admin/seo-pages:
    get:
      operationId: getAdminSeoPages
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All SEO landing pages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SeoPage"
    post:
      operationId: createSeoPage
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeoPageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeoPage"
  /admin/seo-pages/{id}:
    patch:
      operationId: updateSeoPage
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeoPageInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeoPage"
    delete:
      operationId: deleteSeoPage
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Web Push ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: admin-token
  schemas:
...
    SeoPage:
      type: object
      required: [id, title, slug, targetKeyword, content, excerpt, category, tags, status, viewCount, isActive, order, seoTitle, seoDescription, seoKeywords, seoSlug, ogImage, canonicalUrl, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        targetKeyword: { type: string }
        content: { type: string }
        excerpt: { type: string }
        coverImage: { type: ["string", "null"] }
        category: { type: string }
        tags: { type: string }
        status:
          type: string
          enum: [draft, published]
        publishedAt: { type: ["string", "null"] }
        viewCount: { type: integer }
        isActive: { type: boolean }
        order: { type: integer }
        seoTitle: { type: string }
        seoDescription: { type: string }
        seoKeywords: { type: string }
        seoSlug: { type: string }
        ogImage: { type: string }
        canonicalUrl: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
    SeoPageInput:
      type: object
      required: [title]
      properties:
        title: { type: string }
        slug: { type: string }
        targetKeyword: { type: string }
        content: { type: string }
        excerpt: { type: string }
        coverImage: { type: ["string", "null"] }
        category: { type: string }
        tags: { type: string }
        status:
          type: string
          enum: [draft, published]
        publishedAt: { type: ["string", "null"] }
        isActive: { type: boolean }
        order: { type: integer }
        seoTitle: { type: string }
        seoDescription: { type: string }
        seoKeywords: { type: string }
        seoSlug: { type: string }
        ogImage: { type: string }
        canonicalUrl: { type: string }